possible reason for memory dump
I think this function is the culprit Im 90% the return ture was in the
wrong location. This function iteratorate through all xbee connections,
and checks to if it all packets for that connection was recieved. ie
(seq/ maxSeq). It supposed to stay at the last next iteration if it
returns true. So connection 5/10 return trues it will not start at
connection 1 but stay at connection 6 for when it returns. If it gets to
10/10 and no pkt found ie pkt == null it should return false.  I think
it was reading the firt pkt as null and returning true causing some
systems to memory dump.